Added some boundary checks.

+++ poppler/CurlCache.h
@@ -58,7 +58,7 @@ public:
   long int tell();
   int seek(long int offset, int origin);
   size_t read(void * ptr, size_t unitsize, size_t count);
-  void preload(long int start, long int end);
+  void preload(size_t start, size_t end);
   
   void loadChunks(int startBlock, int endBlock);
